@@ -8,15 +8,17 @@
     },
     "dependencies": {
         "aphrodite": "^2.4.0",
-        "firebase": "^9.23.0",
-        "react": "^17.0.2",
+        "firebase": "^10.4.0",
+        "react": "^18.2.0",
         "react-animations": "^1.0.0",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.2.0",
         "react-loading-skeleton": "^3.3.1",
-        "shards-react": "^1.0.3"
+        "react-particles": "^2.12.2",
+        "shards-react": "^1.0.3",
+        "tsparticles-slim": "^2.12.0"
     },
     "devDependencies": {
-        "@vitejs/plugin-react": "^1.0.7",
-        "vite": "^2.7.2"
+        "@vitejs/plugin-react": "^1.3.2",
+        "vite": "^2.9.16"
     }
 }
